Added Derby support.
Separated coupling between JdbcDao and ChannelController so that JdbcDao no longer calls ChannelController to lookup the localChannelId for a channelId, which in turn called JdbcDao again. The map of channelIds to localChannelIds is now stored in JdbcDao.

Donkey's channel storage is now initialized on channel deploy.

Modified channel deploy to check if a ResponseSelector has been defined and if not, create one using the source connector's inbound data type.

Fixed some potential problems with error handling in donkey Channel/DestinationChain.

Changed XmlQuerySource.getQuery() and CachedPreparedStatementSource.getPreparedStatement() so that they return null if there is no query for the given queryId, rather than throwing an exception. This was done since some queries are only defined for specific databases.

Cleaned up donkey's unit tests and ensured that database connections, statements, and result sets are only opened when needed and are closed as soon as possible in finally blocks.

Removed the com.mirth.connect.server.tasks package and the classes within it, which are unused.

Added / cleaned up indexes on donkey tables.

Added debug log output to JdbcDao.

Made the custom meta-data columns case-insensitive by automatically converting them to uppercase.
Issue: MIRTH-2250

+++ donkey/src/test/java/com/mirth/connect/donkey/test/ConnectorTests.java
@@ -61,8 +61,10 @@ public class ConnectorTests {
         String channelId = TestUtils.DEFAULT_CHANNEL_ID;
         String serverId = TestUtils.DEFAULT_SERVER_ID;
 
-        ChannelController.getInstance().deleteAllMessages(channelId);
-
+        if (ChannelController.getInstance().channelExists(channelId)) {
+            ChannelController.getInstance().deleteAllMessages(channelId);
+        }
+        
         TestChannel channel = new TestChannel();
 
         channel.setChannelId(channelId);
